+++ client/src/components/common/AuthFormInput/index.tsx
@@ -1,0 +1,31 @@
+import { forwardRef } from 'react';
+import styles from './AuthFormInput.module.scss';
+import classNames from 'classnames/bind';
+
+const cn = classNames.bind(styles);
+
+interface AuthFormInputProps {
+  placeholder: string;
+  helpText?: string;
+  hasError?: boolean;
+  [key: string]: any;
+}
+
+// TODO: memo
+const AuthFormInput = forwardRef<HTMLInputElement, AuthFormInputProps>(
+  (props, ref) => {
+    const { helpText, hasError, ...others } = props;
+    return (
+      <>
+        <input
+          className={cn('input', { error: hasError })}
+          ref={ref}
+          {...others}
+        />
+        {helpText && <p className={cn('help')}>{helpText}</p>}
+      </>
+    );
+  }
+);
+
+export default AuthFormInput;
