+++ client/src/components/common/AuthFormInputWithErrorMessageWrapper/index.tsx
@@ -1,0 +1,8 @@
+import { PropsWithChildren } from 'react';
+import styles from './AuthFormInputWithErrorMessageWrapper.module.scss';
+
+const AuthFormInputWithErrorMessageWrapper = ({
+  children,
+}: PropsWithChildren) => <div className={styles.wrapper}>{children}</div>;
+
+export default AuthFormInputWithErrorMessageWrapper;
