--- client/src/components/common/FormInputWithErrorMessageWrapper/index.tsx
@@ -1,8 +1,0 @@
-import { PropsWithChildren } from 'react';
-import styles from './FormInputWithErrorMessageWrapper.module.scss';
-
-const FormInputWithErrorMessageWrapper = ({ children }: PropsWithChildren) => (
-  <div className={styles.wrapper}>{children}</div>
-);
-
-export default FormInputWithErrorMessageWrapper;
